requirements.txt
accelerate==1.2.1 ; python_version >= "3.10" and python_version < "3.11"
aiofiles==23.2.1 ; python_version >= "3.10" and python_version < "3.11"
annotated-types==0.7.0 ; python_version >= "3.10" and python_version < "3.11"
anyio==4.8.0 ; python_version >= "3.10" and python_version < "3.11"
attrs==25.3.0 ; python_version >= "3.10" and python_version < "3.11"
certifi==2024.12.14 ; python_version >= "3.10" and python_version < "3.11"
charset-normalizer==3.4.1 ; python_version >= "3.10" and python_version < "3.11"
click==8.1.8 ; python_version >= "3.10" and python_version < "3.11"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.11" and platform_system == "Windows"
docling-core==2.48.1 ; python_version >= "3.10" and python_version < "3.11"
exceptiongroup==1.2.2 ; python_version >= "3.10" and python_version < "3.11"
fastapi==0.115.6 ; python_version >= "3.10" and python_version < "3.11"
ffmpy==0.5.0 ; python_version >= "3.10" and python_version < "3.11"
filelock==3.16.1 ; python_version >= "3.10" and python_version < "3.11"
fsspec==2024.12.0 ; python_version >= "3.10" and python_version < "3.11"
gradio-client==1.5.2 ; python_version >= "3.10" and python_version < "3.11"
gradio==5.9.1 ; python_version >= "3.10" and python_version < "3.11"
h11==0.14.0 ; python_version >= "3.10" and python_version < "3.11"
httpcore==1.0.7 ; python_version >= "3.10" and python_version < "3.11"
httpx==0.28.1 ; python_version >= "3.10" and python_version < "3.11"
huggingface-hub==0.27.1 ; python_version >= "3.10" and python_version < "3.11"
idna==3.10 ; python_version >= "3.10" and python_version < "3.11"
jinja2==3.1.5 ; python_version >= "3.10" and python_version < "3.11"
jsonref==1.1.0 ; python_version >= "3.10" and python_version < "3.11"
jsonschema-specifications==2025.9.1 ; python_version >= "3.10" and python_version < "3.11"
jsonschema==4.25.1 ; python_version >= "3.10" and python_version < "3.11"
latex2mathml==3.78.1 ; python_version >= "3.10" and python_version < "3.11"
markdown-it-py==3.0.0 ; python_version >= "3.10" and python_version < "3.11"
markupsafe==2.1.5 ; python_version >= "3.10" and python_version < "3.11"
mdurl==0.1.2 ; python_version >= "3.10" and python_version < "3.11"
mpmath==1.3.0 ; python_version >= "3.10" and python_version < "3.11"
networkx==3.4.2 ; python_version >= "3.10" and python_version < "3.11"
numpy==2.2.1 ; python_version >= "3.10" and python_version < "3.11"
nvidia-cublas-cu12==12.1.3.1 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10" and python_version < "3.11"
nvidia-cuda-cupti-cu12==12.1.105 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10" and python_version < "3.11"
nvidia-cuda-nvrtc-cu12==12.1.105 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10" and python_version < "3.11"
nvidia-cuda-runtime-cu12==12.1.105 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10" and python_version < "3.11"
nvidia-cudnn-cu12==9.1.0.70 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10" and python_version < "3.11"
nvidia-cufft-cu12==11.0.2.54 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10" and python_version < "3.11"
nvidia-curand-cu12==10.3.2.106 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10" and python_version < "3.11"
nvidia-cusolver-cu12==11.4.5.107 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10" and python_version < "3.11"
nvidia-cusparse-cu12==12.1.0.106 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10" and python_version < "3.11"
nvidia-nccl-cu12==2.20.5 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10" and python_version < "3.11"
nvidia-nvjitlink-cu12==12.6.85 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10" and python_version < "3.11"
nvidia-nvtx-cu12==12.1.105 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10" and python_version < "3.11"
orjson==3.10.13 ; python_version >= "3.10" and python_version < "3.11"
packaging==24.2 ; python_version >= "3.10" and python_version < "3.11"
pandas==2.2.3 ; python_version >= "3.10" and python_version < "3.11"
pillow==11.1.0 ; python_version >= "3.10" and python_version < "3.11"
psutil==5.9.8 ; python_version >= "3.10" and python_version < "3.11"
pydantic-core==2.27.2 ; python_version >= "3.10" and python_version < "3.11"
pydantic==2.10.4 ; python_version >= "3.10" and python_version < "3.11"
pydub==0.25.1 ; python_version >= "3.10" and python_version < "3.11"
pygments==2.19.1 ; python_version >= "3.10" and python_version < "3.11"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "3.11"
python-multipart==0.0.20 ; python_version >= "3.10" and python_version < "3.11"
pytz==2024.2 ; python_version >= "3.10" and python_version < "3.11"
pyyaml==6.0.2 ; python_version >= "3.10" and python_version < "3.11"
referencing==0.36.2 ; python_version >= "3.10" and python_version < "3.11"
regex==2024.11.6 ; python_version >= "3.10" and python_version < "3.11"
requests==2.32.3 ; python_version >= "3.10" and python_version < "3.11"
rich==13.9.4 ; python_version >= "3.10" and python_version < "3.11"
rpds-py==0.27.1 ; python_version >= "3.10" and python_version < "3.11"
ruff==0.8.6 ; python_version >= "3.10" and python_version < "3.11" and sys_platform != "emscripten"
safehttpx==0.1.6 ; python_version >= "3.10" and python_version < "3.11"
safetensors==0.5.1 ; python_version >= "3.10" and python_version < "3.11"
semantic-version==2.10.0 ; python_version >= "3.10" and python_version < "3.11"
shellingham==1.5.4 ; python_version >= "3.10" and python_version < "3.11"
six==1.17.0 ; python_version >= "3.10" and python_version < "3.11"
sniffio==1.3.1 ; python_version >= "3.10" and python_version < "3.11"
spaces==0.30.4 ; python_version >= "3.10" and python_version < "3.11"
starlette==0.41.3 ; python_version >= "3.10" and python_version < "3.11"
sympy==1.13.3 ; python_version >= "3.10" and python_version < "3.11"
tabulate==0.9.0 ; python_version >= "3.10" and python_version < "3.11"
tokenizers==0.21.0 ; python_version >= "3.10" and python_version < "3.11"
tomlkit==0.13.2 ; python_version >= "3.10" and python_version < "3.11"
torch==2.4.0 ; python_version >= "3.10" and python_version < "3.11"
tqdm==4.67.1 ; python_version >= "3.10" and python_version < "3.11"
transformers==4.47.1 ; python_version >= "3.10" and python_version < "3.11"
triton==3.0.0 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version < "3.11" and python_version >= "3.10"
typer==0.15.1 ; python_version >= "3.10" and python_version < "3.11"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "3.11"
tzdata==2024.2 ; python_version >= "3.10" and python_version < "3.11"
urllib3==2.3.0 ; python_version >= "3.10" and python_version < "3.11"
uvicorn==0.34.0 ; python_version >= "3.10" and python_version < "3.11" and sys_platform != "emscripten"
websockets==14.1 ; python_version >= "3.10" and python_version < "3.11"